requirements.txt
numpy==1.14.5
flask
pygame
pyglet
opencv-python
opencv-contrib-python
pybind11
BriCA2
git+https://github.com/TaroSunagawa/oculoenv.git
tensorflow
chainer
torch
torchvision
#oculoenv==0.1.0